pyproject.toml
[project]
name = "github_events"
version = "0.1.0"
description = "Lorem Ipsum"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "matplotlib>=3.10.5",
    "pandas>=2.3.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
